move solvers to separate packages
