+++ hadrons/solver.py
@@ -1,8 +1,8 @@
 from enum import Enum, auto
 from hadrons.cython_files.initial_recombination import single_track_PDEsolver as cython_single_track_PDEsolver
-from hadrons.initial_recombination import single_track_PDEsolver as python_single_track_PDEsolver
-from hadrons.initial_recombination_numba import single_track_PDEsolver as numba_single_track_PDEsolver
-from hadrons.initial_recombination_numba_parallel import single_track_PDEsolver as numba_parallel_single_track_PDEsolver
+from hadrons.python.initial_recombination import single_track_PDEsolver as python_single_track_PDEsolver
+from hadrons.numba.initial_recombination_numba import single_track_PDEsolver as numba_single_track_PDEsolver
+from hadrons.parallel.initial_recombination_numba_parallel import single_track_PDEsolver as numba_parallel_single_track_PDEsolver
 
 class SolverType(Enum):
     CYTHON = auto()
